site/requirements.txt
# TinyTorch Course Dependencies for Site Documentation Builds
# Note: For Binder/Colab environments, see binder/requirements.txt
# Keep synchronized with main requirements.txt

# Core numerical computing
numpy>=1.24.0,<3.0.0
matplotlib>=3.5.0

# Data handling
PyYAML>=6.0

# Rich terminal formatting (for development feedback)
rich>=13.0.0

# Jupyter Book for building documentation
jupyter-book>=0.15.0,<1.0.0

# Jupyter environment
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Sphinx extensions
sphinxcontrib-mermaid>=0.9.2

# Type checking support
typing-extensions>=4.0.0

# For executing TinyTorch code
setuptools>=70.0.0
wheel>=0.42.0

# Optional: for advanced visualizations
# plotly>=5.0.0
# seaborn>=0.11.0
